+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Added
+
+- Field now support CollapseLabel
+
 ### Fixed
 
 - `ButtonGroup`/`ButtonToggle` display issues